requirements.txt
Flask>=2.0
google-generativeai>=0.4.0
python-dotenv>=0.19.0
Pillow>=9.0.0 